requirements.txt
openai
python-dotenv
pypdf2